+++ packages/Icon/src/Check.d.ts
@@ -1,5 +1,0 @@
-export default SvgCheck;
-declare function SvgCheck({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;