+++ packages/Icon/src/Clock.d.ts
@@ -1,5 +1,0 @@
-export default SvgClock;
-declare function SvgClock({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;