+++ packages/Icon/src/Dash.d.ts
@@ -1,5 +1,0 @@
-export default SvgDash;
-declare function SvgDash({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;