+++ packages/Icon/src/EllipsisVertical.d.ts
@@ -1,5 +1,0 @@
-export default SvgEllipsisVertical;
-declare function SvgEllipsisVertical({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;