+++ packages/Icon/src/TimeAndDate.d.ts
@@ -1,5 +1,0 @@
-export default SvgTimeAndDate;
-declare function SvgTimeAndDate({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;