+++ packages/Icon/src/TimesCircle.d.ts
@@ -1,5 +1,0 @@
-export default SvgTimesCircle;
-declare function SvgTimesCircle({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;