+++ packages/Icon/src/Unlink.d.ts
@@ -1,5 +1,0 @@
-export default SvgUnlink;
-declare function SvgUnlink({ title, ...props }: {
-    [x: string]: any;
-    title: any;
-}): JSX.Element;