+++ packages/Icon/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed `title` prop should be optional
+- Removed `d.ts` from `src` folder
+
 ## [0.3.0] - 2020-04-29
 
 ### Removed